Rescues a test utility from fixtures and modifies estimate gas metrics to use standard compiler output.
