+++ tests/utils/ursula.py
@@ -23,15 +23,23 @@ import tempfile
 from cryptography.x509 import Certificate
 from typing import Iterable, List, Optional, Set
 
+from nucypher.characters.lawful import Bob
+from nucypher.crypto.utils import canonical_address_from_umbral_key
+
 from nucypher.blockchain.eth.actors import Staker
 from nucypher.blockchain.eth.interfaces import BlockchainInterface
 from nucypher.characters.lawful import Ursula
 from nucypher.config.characters import UrsulaConfiguration
 from nucypher.crypto.powers import TransactingPower
+from nucypher.policy.collections import WorkOrder, IndisputableEvidence
 from tests.constants import (
     MOCK_URSULA_DB_FILEPATH,
     NUMBER_OF_URSULAS_IN_DEVELOPMENT_NETWORK
 )
+from umbral import pre
+from umbral.curvebn import CurveBN
+from umbral.keys import UmbralPrivateKey
+from umbral.signing import Signer
 
 
 def select_test_port() -> int:
@@ -166,3 +174,51 @@ def start_pytest_ursula_services(ursula: Ursula) -> Certificate:
 
 MOCK_KNOWN_URSULAS_CACHE = dict()
 MOCK_URSULA_STARTING_PORT = 51000  # select_test_port()
+
+
+def _mock_ursula_reencrypts(ursula, corrupt_cfrag: bool = False):
+    delegating_privkey = UmbralPrivateKey.gen_key()
+    _symmetric_key, capsule = pre._encapsulate(delegating_privkey.get_pubkey())
+    signing_privkey = UmbralPrivateKey.gen_key()
+    signing_pubkey = signing_privkey.get_pubkey()
+    signer = Signer(signing_privkey)
+    priv_key_bob = UmbralPrivateKey.gen_key()
+    pub_key_bob = priv_key_bob.get_pubkey()
+    kfrags = pre.generate_kfrags(delegating_privkey=delegating_privkey,
+                                 signer=signer,
+                                 receiving_pubkey=pub_key_bob,
+                                 threshold=2,
+                                 N=4,
+                                 sign_delegating_key=False,
+                                 sign_receiving_key=False)
+    capsule.set_correctness_keys(delegating_privkey.get_pubkey(), pub_key_bob, signing_pubkey)
+
+    ursula_pubkey = ursula.stamp.as_umbral_pubkey()
+
+    alice_address = canonical_address_from_umbral_key(signing_pubkey)
+    blockhash = bytes(32)
+
+    specification = b''.join((bytes(capsule),
+                              bytes(ursula_pubkey),
+                              bytes(ursula.decentralized_identity_evidence),
+                              alice_address,
+                              blockhash))
+
+    bobs_signer = Signer(priv_key_bob)
+    task_signature = bytes(bobs_signer(specification))
+
+    metadata = bytes(ursula.stamp(task_signature))
+
+    cfrag = pre.reencrypt(kfrags[0], capsule, metadata=metadata)
+
+    if corrupt_cfrag:
+        cfrag.proof.bn_sig = CurveBN.gen_rand(capsule.params.curve)
+
+    cfrag_signature = bytes(ursula.stamp(bytes(cfrag)))
+
+    bob = Bob.from_public_keys(verifying_key=pub_key_bob)
+    task = WorkOrder.PRETask(capsule, task_signature, cfrag, cfrag_signature)
+    work_order = WorkOrder(bob, None, alice_address, [task], None, ursula, blockhash)
+
+    evidence = IndisputableEvidence(task, work_order)
+    return evidence
